% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actual_and_forecast.R
\name{actual_and_forecast}
\alias{actual_and_forecast}
\title{Title: actual_and_forecast carbon intensity}
\usage{
actual_and_forecast(date)
}
\arguments{
\item{date}{A character string in the format 'YYYY-MM-DD' representing the date}
}
\value{
A ggplot object representing the trend of carbon intensity, both actual and forecasted.
}
\description{
This function fetches the actual and forecasted carbon intensity data for electricity generation
in Great Britain for a specified date from the National Grid's Carbon Intensity API.
It then plots this data showing the trends over the course of the day.
}
\examples{
actual_and_forecast('2023-01-12')
}
