% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbon_generation.R
\name{carbon_generation}
\alias{carbon_generation}
\title{Get Dataframe of Carbon Generation Mix}
\usage{
carbon_generation(from_date, to_date)
}
\arguments{
\item{from_date}{The starting date of the range.}

\item{to_date}{The ending date of the range.}
}
\value{
A dataframe containing information about the carbon generation mix.
}
\description{
This function retrieves data on carbon generation mix from the specified date range
using the Carbon Intensity API.
}
\examples{
carbon_generation("2022-01-01", "2022-01-15")

}
