% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_CarbonStats.R, R/actual_and_forecast.R,
%   R/carbon_generation.R
\docType{package}
\name{carbonVisR}
\alias{carbonVisR}
\title{Carbon-Intensity Visualization Package}
\description{
This package provides functions for visualizing carbon generation data.
}
\references{
\href{https://carbon-intensity.github.io/api-definitions/#carbon-intensity-api-v2-0-0}{Carbon Intensity API}
}
\seealso{
\code{\link{Get_CarbonStats}}(ZZ)

\code{\link{actual_and_forecast}}(RM)

\code{\link{validate_date_range}}(YC), \code{\link{carbon_generation}}(YC), \code{\link{carbon_generation_visual}}(YC)
}
\author{
Zheng Zhang

Rui Mao

Yahan Cong
}
\keyword{API}
\keyword{carbon}
\keyword{generation}
\keyword{intensity,}
\keyword{mix,}
\keyword{visualization,}
