% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_CarbonStats.R
\name{Get_CarbonStats}
\alias{Get_CarbonStats}
\title{Get Carbon Intensity Statistics}
\usage{
Get_CarbonStats(from, to)
}
\arguments{
\item{from}{Start date in "YYYY-MM-DD" format.}

\item{to}{End date in "YYYY-MM-DD" format.}
}
\value{
A list with two elements: 'data' (a data frame) and 'plot' (a ggplot object).
}
\description{
This function retrieves carbon intensity statistics from the Carbon Intensity API
for a specified date range. It returns a list containing two elements: 'data' (a
data frame) and 'plot' (a ggplot object) for visualizing the intensity trends.
}
\details{
The function queries the Carbon Intensity API for each day within the specified
date range and retrieves maximum, average, and minimum intensity values. The result
is a data frame containing date-wise intensity statistics, and a ggplot object for
visualizing the intensity trends over time.
}
\examples{
Get_CarbonStats("2023-01-20", "2023-01-30")
Get_CarbonStats("2023-01-20", "2023-01-30")$data
Get_CarbonStats("2023-01-20", "2023-01-30")$plot


}
