% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbon_generation.R
\docType{package}
\name{carbonVisR}
\alias{-package}
\alias{carbonVisR}
\title{Carbon Visualization Package}
\description{
This package provides functions for visualizing carbon generation data.
}
\author{
Yahan
}
